Remove basicaer (#546)
* removed legacy ci

* Delete .github/workflows/tests.yml

* naming and 0.46 test

* rename

* update qiskit device

* dep warnings

* pylint

* removed devices

* pylint

* pylint

* docs fix

* setup.py changes

* delete

* revert

* reqs change

* setup

* change to reqs to match ci

* removed a test

* pylint

* put ibmq.rst back

* delete ibmq

* remove ibmq

* deletion

* codecov

* lint

* changes to tests

* Revert "Remove devices that will not be supported in the new release" (#544)

* docs

* remove basicaer

* reqs to build docs

* docs

* pylint

* tests fix

* path change

* Changelog and doc

* changelogs

* Update CHANGELOG.md

* deleted a test file

* removed error

* basic sim pylint

* Update CHANGELOG.md


* remove ifelse block

* pylint

---------
