+++ CHANGELOG.md
@@ -13,6 +13,9 @@
 * The test suite no longer runs for Qiskit versions below 0.46.
   [(#536)](https://github.com/PennyLaneAI/pennylane-qiskit/pull/536)
 
+* The ``qiskit.basicaer`` device has been removed because it is not supported for versions of Qiskit above 0.46.
+  [(#546)](https://github.com/PennyLaneAI/pennylane-qiskit/pull/546)
+
 ### Deprecations 👋
 
 ### Documentation 📝
