+++ doc/requirements.txt
@@ -7,10 +7,10 @@ pennylane==0.34
 pybind11==2.11.1
 pygments==2.17.2
 pygments-github-lexers==0.0.5
-qiskit==0.45.3
-qiskit-aer==0.13.3
-qiskit-ibm-runtime==0.20.0
-qiskit-ibm-provider==0.10.0
+qiskit==1.0.2
+qiskit-aer==0.14.1
+qiskit-ibm-runtime==0.23.0
+qiskit-ibm-provider==0.11.0
 sphinxcontrib-bibtex==2.6.2
 sphinx-automodapi==0.17.0
 pennylane-sphinx-theme
