+++ tests/conftest.py
@@ -18,12 +18,10 @@ This module contains some configuration for PennyLane IBMQ devices.
 import os
 import pytest
 import numpy as np
-import qiskit
 
 import pennylane as qml
-from semantic_version import Version
 from qiskit_ibm_provider import IBMProvider
-from pennylane_qiskit import AerDevice, BasicAerDevice, BasicSimulatorDevice
+from pennylane_qiskit import AerDevice, BasicSimulatorDevice
 
 # pylint: disable=protected-access, unused-argument, redefined-outer-name
 
@@ -40,22 +38,14 @@ U2 = np.array([[0, 1, 1, 1], [1, 0, 1, -1], [1, -1, 0, 1], [1, 1, -1, 0]]) / np.
 A = np.array([[1.02789352, 1.61296440 - 0.3498192j], [1.61296440 + 0.3498192j, 1.23920938 + 0j]])
 
 
-if Version(qiskit.__version__) < Version("1.0.0"):
-    test_devices = [AerDevice, BasicAerDevice]
-    hw_backends = ["qasm_simulator", "aer_simulator"]
-    state_backends = [
-        "statevector_simulator",
-        "unitary_simulator",
-    ]
-else:
-    test_devices = [AerDevice, BasicSimulatorDevice]
-    hw_backends = ["qasm_simulator", "aer_simulator", "basic_simulator"]
-    state_backends = [
-        "statevector_simulator",
-        "unitary_simulator",
-        "aer_simulator_statevector",
-        "aer_simulator_unitary",
-    ]
+test_devices = [AerDevice, BasicSimulatorDevice]
+hw_backends = ["qasm_simulator", "aer_simulator", "basic_simulator"]
+state_backends = [
+    "statevector_simulator",
+    "unitary_simulator",
+    "aer_simulator_statevector",
+    "aer_simulator_unitary",
+]
 
 
 @pytest.fixture
